backend/backup-20260108/youtube-heat-map/requirements.txt
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
